Refactor internals so we can expose a "multi" API
Our goal is to break apart all operations into "prepare" and "process"
pairs so that we can use an external event loop for all I/O instead of
cURL's blocking easy mode.

We prepare for that here by cleaning up the request/response processing
internals so that they'll be easier to break up in a future commit.